create_table_backup {aws.dynamodb}	R Documentation

Backup a Dynamo DB Table

Description

Create a backup of a Dynamo DB table, or restore a table from a backup

Usage

create_table_backup(table, backup, ...)

restore_table_backup(table, backup, ...)

Arguments

table
A character string specifying the table name, or an object of class “aws_dynamodb_table”.

backup
For creating: a character string specifying the backup table name. For restoring: a backup table ARN.

...
Additional arguments passed to dynamoHTTP.

Value

A list.

References

API Guide: CreateBackup API Guide: RestoreTableFromBackup

See Also

list_backups

[Package aws.dynamodb version 0.1.2 Index]