dynamoHTTP {aws.dynamodb}	R Documentation

Execute Dynamo DB API Request

Description

This is the workhorse function to execute calls to the Dynamo DB API.

Usage

dynamoHTTP(
  verb = "GET",
  headers = list(),
  body = NULL,
  target,
  verbose = getOption("verbose", FALSE),
  region = Sys.getenv("AWS_DEFAULT_REGION", "us-east-1"),
  key = NULL,
  secret = NULL,
  session_token = NULL,
  url = Sys.getenv("AWS_LOCAL_ENDPOINT"),
  ...
)

Arguments

verb
A character string specifying an HTTP verb for the operation.

headers
A list of headers to pass to the HTTP request.

body
A list of body (JSON) arguments.

target
An API endpoint.

verbose
A logical indicating whether to be verbose. Default is given by options("verbose").

region
A character string containing an AWS region. If missing, the default “us-east-1” is used.

key
A character string containing an AWS Access Key ID. See locate_credentials.

secret
A character string containing an AWS Secret Access Key. See locate_credentials.

session_token
Optionally, a character string containing an AWS temporary Session Token. See locate_credentials.

...
Additional arguments passed to GET.

Details

This function constructs and signs an Dynamo DB API request and returns the results thereof, or relevant debugging information in the case of error.

Value

If successful, a named list. Otherwise, a data structure of class “aws-error” containing any error message(s) from AWS and information about the request attempt.

Author(s)

Thomas J. Leeper

[Package aws.dynamodb version 0.1.2 Index]