list_table_backups {aws.dynamodb}	R Documentation

List Dynamo DB Backups

Description

Retreive a paginated list of backup tables

Usage

list_table_backups(marker = NULL, n = 100, ...)

Arguments

marker
A character string containing the first backup table to return (contained as an attribute in a previous call to this function)

n
A number between 1 and 100 specifying the total number of tables to return

...
Additional arguments passed to dynamoHTTP.

Value

A list of backup tables.

References

API Guide: ListBackups

[Package aws.dynamodb version 0.1.2 Index]