update_item {aws.dynamodb}	R Documentation

Update Dynamo DB Item

Description

Update an item from a Dynamo DB tables

Usage

update_item(table, key, item, return_value = c("NONE", "ALL_OLD"), ...)

Arguments

table
A character string specifying the table name, or an object of class “aws_dynamodb_table”.

key
A list of key-value pairs. If the table only has a primary key, this should specify the primary key attribute name and value for the desired item. If a composite primary key is used, then both attribute names and values must be specified.

item
A list of key-value pairs.

...
Additional arguments passed to dynamoHTTP.

Value

A list.

References

API Guide: UpdateItem

Examples

## Not run: 
  tab <- create_table(
    table = "Music",
    attributes = list(Artist = "S"),
    primary_key = "Artist"
  )
  
  # put item
  put_item("Music", list(Artist = "No One You Know", SongTitle = "Call Me Yesterday"))

  # get item
  get_item("Music", list(Artist = "No One You Know"))
  
  # update item
  update_item("Music", key = list(Artist = "No One You Know", Songtitle = "Call Me Today"))
  
  # get item
  get_item("Music", list(Artist = "No One You Know"))

  # cleanup
  delete_table(tab)

## End(Not run)

[Package aws.dynamodb version 0.1.2 Index]